moved functionality of chat display to it's own component, and rendering this in the index page.
