+++ src/pages/index.tsx
@@ -11,65 +11,66 @@ import { api } from '~/utils/api';
 import { useRef, useState } from 'react';
 import React from 'react';
 import { UserButton, useUser } from '@clerk/nextjs';
+import ChatInterface from '~/components/Chat/ChatInterface';
 
 export default function Home() {
   const mobileScreen = useMediaQuery('(max-width: 480px)');
 
-  const [chatItems, setChatItems] = useState<ChatItem[]>([]);
-  const [waiting, setWaiting] = useState<boolean>(false);
+  // const [chatItems, setChatItems] = useState<ChatItem[]>([]);
+  // const [waiting, setWaiting] = useState<boolean>(false);
 
-  const generatedTextMutation = api.ai.generateText.useMutation({
-    onSuccess: (data) => {
-      setChatItems([
-        ...chatItems,
-        {
-          content: data.generatedText,
-          author: 'AI',
-        },
-      ]);
-    },
+  // const generatedTextMutation = api.ai.generateText.useMutation({
+  //   onSuccess: (data) => {
+  //     setChatItems([
+  //       ...chatItems,
+  //       {
+  //         content: data.generatedText,
+  //         author: 'AI',
+  //       },
+  //     ]);
+  //   },
 
-    onError: (error) => {
-      setChatItems([
-        ...chatItems,
-        {
-          content: error.message ?? 'An error occurred',
-          author: 'AI',
-          isError: true,
-        },
-      ]);
-    },
+  //   onError: (error) => {
+  //     setChatItems([
+  //       ...chatItems,
+  //       {
+  //         content: error.message ?? 'An error occurred',
+  //         author: 'AI',
+  //         isError: true,
+  //       },
+  //     ]);
+  //   },
 
-    onSettled: () => {
-      setWaiting(false);
-    },
-  });
+  //   onSettled: () => {
+  //     setWaiting(false);
+  //   },
+  // });
 
-  const resetMutation = api.ai.reset.useMutation();
+  // const resetMutation = api.ai.reset.useMutation();
 
-  const handleUpdate = (prompt: string) => {
-    setWaiting(true);
+  // const handleUpdate = (prompt: string) => {
+  //   setWaiting(true);
 
-    setChatItems([
-      ...chatItems,
-      {
-        content: prompt.replace(/\n/g, '\n\n'),
-        author: 'User',
-      },
-    ]);
+  //   setChatItems([
+  //     ...chatItems,
+  //     {
+  //       content: prompt.replace(/\n/g, '\n\n'),
+  //       author: 'User',
+  //     },
+  //   ]);
 
-    console.log('User sent a message:', prompt);
-    console.log('Before calling mutate:', chatItems);
-    console.log('Prompt value:', prompt);
+  //   console.log('User sent a message:', prompt);
+  //   console.log('Before calling mutate:', chatItems);
+  //   console.log('Prompt value:', prompt);
 
-    generatedTextMutation.mutate({ prompt });
+  //   generatedTextMutation.mutate({ prompt });
 
-    console.log('After calling mutate:', chatItems);
-  };
-  const handleReset = () => {
-    setChatItems([]);
-    resetMutation.mutate();
-  };
+  //   console.log('After calling mutate:', chatItems);
+  // };
+  // const handleReset = () => {
+  //   setChatItems([]);
+  //   resetMutation.mutate();
+  // };
 
   const { isLoaded, user } = useUser();
   return (
@@ -84,14 +85,15 @@ export default function Home() {
       )}
       {isLoaded && user && (
         <Box>
-          {mobileScreen ? <HeaderMobile onReset={handleReset} /> : <Header />}
-          {mobileScreen ? null : <Sidebar onReset={handleReset} />}
+          {/* {mobileScreen ? <HeaderMobile onReset={handleReset} /> : <Header />} */}
+          {/* {mobileScreen ? null : <Sidebar onReset={handleReset} />}
           <ChatContent
             chatItems={chatItems}
             onReset={handleReset}
             loading={waiting}
           />
-          <ChatInput onUpdate={handleUpdate} waiting={waiting} />
+          <ChatInput onUpdate={handleUpdate} waiting={waiting} /> */}
+          <ChatInterface />
           <Footer />
         </Box>
       )}
